# src/api/openapi.yaml
openapi: 3.0.3
info:
  description: "# Introduction\n\nThe Bridge API serves as a REST interface which\
    \ is able to fetch content, product and category information from any shop backend\
    \ and to display them in reports in the FirstSpirit ContentCreator.\n\nIn order\
    \ to connect the Bridge API with a given shop backend a Bridge is needed.\nIt\
    \ acts as a microservice between the shop backend and FirstSpirit.\nFurther information\
    \ about how to implement and use a Bridge can be found in the official [documentation](https://docs.e-spirit.com/ecom/fsconnect-com/FirstSpirit_Connect_for_Commerce_Documentation_EN.html).\n\
    \n# Authentication\n\nThe Bridge API supports HTTP Basic Authentication in order\
    \ to secure the API endpoints.\nIn this case it expects a specific HTTP header\
    \ where the client credentials are encoded in base64. Client credentials consist\
    \ of a custom username and password.\n\nExample:\n```\ncurl -u username:password\
    \ -i -H 'Accept:application/json' http://example.com\n```\n\nThe Bridge itself\
    \ requires Basic Auth to provide a secure data exchange with FirstSpirit.\n\n\
    # Errors\n\nSuccess or failure of an API request is indicated by HTTP status codes.\n\
    Status codes starting with ```2xx``` indicate success whereas codes starting with\
    \ ```4xx``` or ```5xx``` indicate that an error occured.\n\n| Status Code    \
...
    \ an internal server error.                        |\n"
  license:
...
  x-logo:
...
servers:
...
security:
- basicAuth: []
...
paths:
  /categories:
    get:
      description: "Returns all available categories from the shop backend in the\
        \ specified language.\n\nIf the specified language is not defined for a category\
        \ or none was set, it displays all categories in the master language by default."
      operationId: categoriesGet
      parameters:
      - description: Filter by parentId. It returns all related categories and subcategories
          from the category tree.
        in: query
        name: parentId
        schema:
          type: string
      - $ref: '#/components/parameters/lang'
      - $ref: '#/components/parameters/page'
      responses:
        "200":
          content:
            application/json:
              examples:
                Categories:
                  $ref: '#/components/examples/Categories'
                  description: Categories
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Labelled'
          description: Categories
          headers:
            X-HasNext:
              $ref: '#/components/headers/hasNext'
              style: simple
            X-Total:
              $ref: '#/components/headers/total'
              style: simple
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get Categories
      tags:
      - Categories
...
  /categories/ids/{categoryIds}:
    get:
      description: Returns a list of categories from the shop backend that match the
        given categoryIds.
      operationId: categoriesCategoryIdsGet
      parameters:
      - $ref: '#/components/parameters/categoryIds'
      - $ref: '#/components/parameters/lang'
      responses:
        "200":
          content:
            application/json:
              examples:
                Category List:
                  $ref: '#/components/examples/CategoryListById'
                  description: Category List
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Labelled'
          description: List of Categories or `null` if it does not exist.
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get list of Categories by their IDs.
      tags:
      - Categories
  /categories/tree:
    get:
      description: "Returns the category tree in the specified language.\n\nIf the\
        \ specified language is not defined for a category or none was set, it displays\
        \ all categories in the master language by default."
      operationId: categoryTreeGet
      parameters:
      - description: Filter by parentId
        in: query
        name: parentId
        schema:
          type: string
      - $ref: '#/components/parameters/lang'
      responses:
        "200":
          content:
            application/json:
              examples:
                Category Tree:
                  $ref: '#/components/examples/CategoryTree'
                  description: Category Tree
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
          description: Category Tree
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get Category Tree
      tags:
      - Categories
    head:
      description: "Checks if the category tree endpoint is available. It is activated\
        \ by default.\n\nIf it is not available, the dropdown for the category tree\
        \ in the ContentCreator is not displayed in both the category and the product\
        \ report."
      operationId: categoryTreeHead
      responses:
        "200":
          description: Category tree endpoint exists
        "404":
          description: Category tree endpoint doesn't exist
      security:
      - basicAuth: []
      summary: Support Category Tree
      tags:
      - Categories
  /categories/{categoryIds}:
    get:
      deprecated: true
      description: "Deprecated: Please use /categories/ids/{categoryIds} instead.\
        \ \n Returns a list of categories from the shop backend that match the given\
        \ categoryIds."
      operationId: categoriesCategoryIdsGetOld
      parameters:
      - $ref: '#/components/parameters/categoryIds'
      - $ref: '#/components/parameters/lang'
      responses:
        "200":
          content:
            application/json:
              examples:
                Category List:
                  $ref: '#/components/examples/CategoryListById'
                  description: Category List
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Labelled'
          description: List of Categories or `null` if it does not exist.
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get list of Categories by their IDs.
      tags:
      - Categories
  /contentpages:
    get:
      description: Returns a pageable list of Content Pages which can be filtered
        using the query parameters listed below.
      operationId: contentPagesGet
      parameters:
      - $ref: '#/components/parameters/query'
      - $ref: '#/components/parameters/lang'
      - $ref: '#/components/parameters/page'
      responses:
        "200":
          content:
            application/json:
              examples:
                Content Page List:
                  $ref: '#/components/examples/ContentPageList'
                  description: Content Page List
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Extracted'
          description: List of Content Pages
          headers:
            X-HasNext:
              $ref: '#/components/headers/hasNext'
              style: simple
            X-Total:
              $ref: '#/components/headers/total'
              style: simple
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get a list of Content Pages
      tags:
      - ContentPages
    head:
      description: |-
        Checks if the Content Report exists in the FirstSpirit module.

        The Content Report is used to display existing content pages from the shop in FirstSpirit. By default the Content Report is activated.
      operationId: contentPagesHead
      responses:
        "200":
          description: Content Report exists
        "404":
          description: Content Report doesn't exist
      security:
      - basicAuth: []
      summary: Support Content Report
      tags:
      - ContentPages
    post:
      description: Create a new Content Page in the storefront and its page reference
        in FirstSpirit.
      operationId: contentPagesPost
      parameters:
      - $ref: '#/components/parameters/lang'
      requestBody:
        content:
          application/json:
            examples:
              Content Page:
                $ref: '#/components/examples/ContentPage'
                description: Content Page
            schema:
              $ref: '#/components/schemas/ContentPage'
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                Content Page:
                  $ref: '#/components/examples/ContentPageResponse'
                  description: Content Page
              schema:
                $ref: '#/components/schemas/IdProviding'
          description: Identifier Object
        "400":
          description: Bad Request. Missing or incorrect parameters
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Create Content Page
      tags:
      - ContentPages
  /contentpages/ids/{contentIds}:
    get:
      description: Returns a list of Content Pages which match the given contentIds.
      operationId: contentPagesContentIdsGet
      parameters:
      - $ref: '#/components/parameters/contentIds'
      - $ref: '#/components/parameters/lang'
      responses:
        "200":
          content:
            application/json:
              examples:
                Content Page List:
                  $ref: '#/components/examples/ContentPageListById'
                  description: Content Page List
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Extracted'
          description: List of Content Pages or `null` if it does not exist
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get list of Content Pages by IDs
      tags:
      - ContentPages
  /contentpages/{contentId}:
    delete:
      description: Remove a Content Page from the Storefront and its page reference
        in FirstSpirit.
      operationId: contentPagesContentIdDelete
      parameters:
      - description: Delete content page by contentId
        in: path
        name: contentId
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/lang'
      responses:
        "204":
          description: Deletion successful
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Delete Content Page
      tags:
      - ContentPages
    put:
      description: Edit information of a Content Page. Changes are reflected in both
        the storefront and in the page reference in FirstSpirit.
      operationId: contentPagesContentIdPut
      parameters:
      - description: Update content page by contentId
        in: path
        name: contentId
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/lang'
      requestBody:
        content:
          application/json:
            examples:
              Content Page:
                $ref: '#/components/examples/ContentPage'
                description: Content Page
            schema:
              $ref: '#/components/schemas/ContentPage'
        required: true
      responses:
        "201":
          description: Update successful
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          description: Content Page not found
      security:
      - basicAuth: []
      summary: Update Content Page
      tags:
      - ContentPages
...
  /products:
    get:
      description: Returns a pageable list of products from the shop backend which
        can be filtered using the query parameters listed below.
      operationId: productsGet
      parameters:
      - description: Filter by categoryId
        in: query
        name: categoryId
        schema:
          type: string
      - $ref: '#/components/parameters/query'
      - $ref: '#/components/parameters/lang'
      - $ref: '#/components/parameters/page'
      responses:
        "200":
          content:
            application/json:
              examples:
                Product List:
                  $ref: '#/components/examples/ProductList'
                  description: Product List
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
          description: List of Products
          headers:
            X-HasNext:
              $ref: '#/components/headers/hasNext'
              style: simple
            X-Total:
              $ref: '#/components/headers/total'
              style: simple
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get list of Products
      tags:
      - Products
...
  /products/ids/{productIds}:
    get:
      description: Returns a list of products from the shop backend which match the
        given productIds.
      operationId: productsProductIdsGet
      parameters:
      - $ref: '#/components/parameters/productIds'
      - $ref: '#/components/parameters/lang'
      responses:
        "200":
          content:
            application/json:
              examples:
                Product List:
                  $ref: '#/components/examples/ProductListById'
                  description: Product List
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
          description: List of Products or `null` if it does not exist
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get list of Products by IDs
      tags:
      - Products
  /products/{productIds}:
    get:
      deprecated: true
      description: "Deprecated: Please use /products/ids/{productIds} instead. \n\
        \ Returns a list of products from the shop backend which match the given productIds."
      operationId: productsProductIdsGetOld
      parameters:
      - $ref: '#/components/parameters/productIds'
      - $ref: '#/components/parameters/lang'
      responses:
        "200":
          content:
            application/json:
              examples:
                Product List:
                  $ref: '#/components/examples/ProductListById'
                  description: Product List
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
          description: List of Products or `null` if it does not exist.
        "401":
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - basicAuth: []
      summary: Get list of Products by IDs
      tags:
      - Products
...
components:
  examples:
    Categories:
      summary: Categories
      value:
      - id: category-1
        label: Electronics
      - id: category-2
        label: Cameras
      - id: category-3
        label: Accessories
      - id: category-4
        label: Headphones
    CategoryListById:
      summary: Category List
      value:
      - id: cat-1
        label: Electronics
      - id: cat-2
        label: Cameras
    CategoryTree:
      summary: Category Tree
      value:
        children:
        - id: category-1-1
          label: Category 1.1
        - children:
          - id: category-1-2-1
            label: Category 1.2.1
          id: category-1-2
          label: Category 1.1
        id: category-1
        label: Category 1
    ContentPage:
      summary: Content Page
      value:
        label: Homepage
        nextSiblingId: null
        pageUid: uid
        parentId: null
        template: content
        visible: true
    ContentPageList:
      summary: Content Page List
      value:
      - extract: /homepage/
        id: homepage
        label: Homepage
      - extract: /my-custom-page/
        id: my-custom-page
        label: My Custom Page
      - extract: /my-custom-page-x/
        id: my-custom-page-x
        label: My Custom Page X
      - extract: /landingpage/
        id: landingpage
        label: Landing Page
    ContentPageListById:
      summary: Content Page List By Id
      value:
      - extract: /homepage/
        id: homepage
        label: Homepage
      - extract: /my-custom-page/
        id: my-custom-page
        label: My Custom Page
    ContentPageResponse:
      summary: Content Page
      value:
        id: content-3
...
    ProductList:
      summary: Product List
      value:
      - extract: Camera
        id: prod-1
        image: https://mycommerce/prod-1/540/960
        label: Sony Legria
        thumbnail: https://mycommerce/prod-1/160/120
      - extract: Camera
        id: prod-2
        image: https://mycommerce/prod-2/540/960
        label: Canon EOS
        thumbnail: https://mycommerce/prod-2/160/120
      - extract: Camera
        id: prod-3
        image: https://mycommerce/prod-3/540/960
        label: PowerShot A480
        thumbnail: https://mycommerce/prod-3/160/120
      - extract: Camera
        id: prod-4
        image: https://mycommerce/prod-4/540/960
        label: Gigashot K80H
        thumbnail: https://mycommerce/prod-4/160/120
    ProductListById:
      summary: Product List
      value:
      - extract: Camera
        id: prod-1
        image: https://mycommerce/prod-1/540/960
        label: Sony Legria
        thumbnail: https://mycommerce/prod-1/160/120
      - extract: Camera
        id: prod-2
        image: https://mycommerce/prod-2/540/960
        label: Canon EOS
        thumbnail: https://mycommerce/prod-2/160/120
...
  headers:
    hasNext:
      description: "Is `true`, if one more pages exist"
      schema:
        type: boolean
    total:
      description: Number of total items
      schema:
        type: integer
  parameters:
    categoryIds:
      description: Get one or more categories by their categoryIds (comma-separated).
      example:
      - ""
      in: path
      name: categoryIds
      required: true
      schema:
        type: array
        items:
          type: string
    contentIds:
      description: Get one or more content pages by contentIds (comma-separated)
      example:
      - ""
      in: path
      name: contentIds
      required: true
      schema:
        type: array
        items:
          type: string
...
    lang:
      description: The language to localize the label
      in: query
      name: lang
      required: false
      schema:
        type: string
    page:
      description: Specific result page
      in: query
      name: page
      schema:
        type: integer
        default: 1
        minimum: 1
    productIds:
      description: Get one or more products by productIds (comma-separated)
      example:
      - ""
      in: path
      name: productIds
      required: true
      schema:
        type: array
        items:
          type: string
    query:
      description: Fulltext search query string
      in: query
      name: q
      schema:
        type: string
...
  responses:
...
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  schemas:
    Category:
      type: object
      properties:
        children:
          type: array
          description: Nested categories containing an id and a label
          items:
            $ref: '#/components/schemas/Category'
        id:
          type: string
          description: "Unique Identifier. Commas, backslashes or whitespace characters\
            \ are forbidden"
          pattern: " ^[^,\\\\\\s]+$"
        label:
          type: string
          description: Localized label
      required:
      - id
      - label
...
    ContentPage:
      type: object
      properties:
        label:
          type: string
          description: Localized label
        nextSiblingId:
          type: string
          description: Content Id of the next sibling page
        pageUid:
          type: string
          description: The uid of the page used in FirstSpirit.
        parentId:
          type: string
          description: Content Id of the parent page
        template:
          type: string
          description: The name of the template used in FirstSpirit.
        visible:
          type: boolean
          description: Defines visibility of a page.
      required:
      - label
      - template
    Extracted:
      type: object
      properties:
        extract:
          type: string
          description: Extract. Representative text for the text extract.
        id:
          type: string
          description: "Unique Identifier. Commas, backslashes or whitespace characters\
            \ are forbidden"
          pattern: " ^[^,\\\\\\s]+$"
        label:
          type: string
          description: Localized label
      required:
      - id
      - label
    IdProviding:
      type: object
      properties:
        id:
          type: string
          description: "Unique Identifier. Commas, backslashes or whitespace characters\
            \ are forbidden"
          pattern: " ^[^,\\\\\\s]+$"
      required:
      - id
...
    Labelled:
      type: object
      properties:
        id:
          type: string
          description: "Unique Identifier. Commas, backslashes or whitespace characters\
            \ are forbidden"
          pattern: " ^[^,\\\\\\s]+$"
        label:
          type: string
          description: Localized label
      required:
      - id
      - label
    Product:
      type: object
      properties:
        extract:
          type: string
          description: Extract. Representative text for the text extract.
        id:
          type: string
          description: "Unique Identifier. Commas, backslashes or whitespace characters\
            \ are forbidden"
          pattern: " ^[^,\\\\\\s]+$"
        image:
          type: string
          format: url
          description: Image URL used for the flyout. Image will be scaled down if
            width exceeds 30vw.
        label:
          type: string
          description: Localized label
        thumbnail:
          type: string
          format: url
          description: Thumbnail URL. Thumbnail will be scaled down if size exceeds
            100x100 pixels.
      required:
      - id
      - label
...
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http